Move command into common
